+++ format.R
@@ -1,0 +1,3 @@
+library(formatR)
+
+tidy_dir(path = "src", recursive = TRUE)