+++ system-setup.sh
@@ -1,0 +1,17 @@
+# /bin/bash
+
+# Install make
+apt-get install -y make
+
+# Install Python venv
+apt-get install -y python3-venv
+
+# Install R
+apt-key adv --keyserver keyserver.ubuntu.com --recv-keys E298A3A825C0D65DFD57CBB651716619E084DAB9
+add-apt-repository "deb https://cloud.r-project.org/bin/linux/ubuntu $(lsb_release -cs)-cran40/"
+apt-get install -y r-base r-base-core r-recommended
+
+# Install Docker
+curl -fsSL https://get.docker.com -o get-docker.sh
+sudo sh get-docker.sh
+rm get-docker.sh