+++ CHANGELOG.rst
@@ -9,3 +9,5 @@ Unreleased
 - Automatically upload local code changes (Issue #7)
 - Use one dask task per study trail (Issue #13)
 - Support remote cluster execution (Issue #6)
+- Improve local execution setup
+- CSL integration via R bridge